# Conflicts:
#	core/src/com/bombcorps/game/view/RoomScreen.java

+++ core/src/com/bombcorps/game/model/Bonus.java
@@ -28,12 +28,12 @@ public class Bonus {
         ACTIVATED, GROUNDED
     }
 
-    public Bonus(int mapWidth){
+    public Bonus(float mapWidth){
         init(mapWidth);
 //        initType();
     }
 
-    public void init(int mapWidth){
+    public void init(float mapWidth){
         state = STATE.ACTIVATED;
         origin = new Vector2();
         scale = new Vector2();
@@ -44,14 +44,7 @@ public class Bonus {
         initPosition(mapWidth);
     }
 
-//    public void initType(){
-//        if(Math.random() > 0.5)
-//            this.type = TYPE.ADDENDURANCE;
-//        else
-//            this.type = TYPE.ADDHEALTH;
-//    }
-
-    public void initPosition(int mapWidth){
+    public void initPosition(float mapWidth){
         position = new Vector2();
         position.x = (float)Math.random() * (mapWidth - dimension.x);
     }
@@ -90,9 +83,6 @@ public class Bonus {
                 origin.x, origin.y, dimension.x, dimension.y,
                 scale.x, scale.y,0);
     }
-
-//    public void
-
 
     public void setState(int input){
         switch (input){
